requirements.txt
# requirements.txt
openai>=1.0.0
google-cloud-speech>=2.20.0
google-cloud-texttospeech>=2.14.0
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.24.0
pyaudio>=0.2.13
python-dotenv>=1.0.0